Use int in Chainer v1

+++ cupy/core/core.pyx
@@ -2202,8 +2202,8 @@ cdef _concatenate_kernel = ElementwiseKernel(
     axis_ind -= cum_sizes[left];
     char* ptr = reinterpret_cast<char*>(x[array_ind]);
     for (int j = ndim - 1; j >= 0; --j) {
-      int ind[] = {array_ind, j};
-      int offset;
+      ptrdiff_t ind[] = {array_ind, j};
+      ptrdiff_t offset;
       if (j == axis) {
         offset = axis_ind;
       } else {
@@ -2312,8 +2312,8 @@ cdef _take_kernel = ElementwiseKernel(
       S wrap_indices = indices % index_range;
       if (wrap_indices < 0) wrap_indices += index_range;
 
-      int li = i / (rdim * cdim);
-      int ri = i % rdim;
+      ptrdiff_t li = i / (rdim * cdim);
+      ptrdiff_t ri = i % rdim;
       out = a[(li * adim + wrap_indices) * rdim + ri];
     ''',
     'cupy_take')
@@ -2359,8 +2359,8 @@ cdef _scatter_update_kernel = ElementwiseKernel(
     '''
       S wrap_indices = indices % adim;
       if (wrap_indices < 0) wrap_indices += adim;
-      int li = i / (rdim * cdim);
-      int ri = i % rdim;
+      ptrdiff_t li = i / (rdim * cdim);
+      ptrdiff_t ri = i % rdim;
       a[(li * adim + wrap_indices) * rdim + ri] = v;
     ''',
     'cupy_scatter_update')
@@ -2372,8 +2372,8 @@ cdef _scatter_add_kernel = ElementwiseKernel(
     '''
       S wrap_indices = indices % adim;
       if (wrap_indices < 0) wrap_indices += adim;
-      int li = i / (rdim * cdim);
-      int ri = i % rdim;
+      ptrdiff_t li = i / (rdim * cdim);
+      ptrdiff_t ri = i % rdim;
       atomicAdd(&a[(li * adim + wrap_indices) * rdim + ri], v[i]);
     ''',
     'cupy_scatter_add')
@@ -3656,7 +3656,7 @@ def _nonzero_1d_kernel(src_dtype, index_dtype):
         const CArray<${index_dtype}, 1> scaned_index,
         CArray<${index_dtype}, 1> dst){
         int thid = blockIdx.x * blockDim.x + threadIdx.x;
-        int n = src.size();
+        ptrdiff_t n = src.size();
         if (thid < n){
             if (src[thid] != 0){
                 dst[scaned_index[thid] - 1] = thid;
